Refactor client.js to remove Netflix dependency
This updates the repro case for the issue so that it doesn't depend on
Netflix internal libraries.

While working on this it became apparent that the issue seems related
to the `round_robin` load balancer option combined with our pattern
of having multiple clients instantiated for the same server (for
client rotation purposes).